post

+++ app/Http/Controllers/TasksController.php
@@ -16,14 +16,22 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $tasks = Task::all();
+        {
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
     }
-    
     // omission
     
     /**
@@ -58,10 +66,14 @@ class TasksController extends Controller
         $task->status = $request->status;    // add
         $task->content = $request->content;
         $task->save();
+        
+        $request->user()->microposts()->create([
+            'content' => $request->content,
+        ]);
 
 
-        return redirect('/');
-    }
+        return redirect()->back();
+        }
 
     /**
      * Display the specified resource.
@@ -123,10 +135,13 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
-        $task = Task::find($id);
-        $task->delete();
+        {
+        $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
+
+        return redirect()->back();
     }
 }
